+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.0
+
+- BREAKING CHANGE: pushover's Notify method now requires a context when being called
+- logger: add HandleErr method
+- linting and other cleanup
+
 # 0.3.0
 
 - constant RACHIO_CONFIG_FILE changed to PROJECT_CONFIG_FILE - note this is somewhat of a breaking change